refactor(gui): extract shared webview preload builder
